Add basic mailserver role
This build a mailserver running postfix, dovecot, rspamd, and sogo. It
works with applied SELinux rules and provides a nice, exchange-ready
webinterface.

it uses traefik as generator for the certificates and dumps them for the
local setup. Maybe needs more harding and splitting up into multiple
roles, but for now it's fine and works well.