mailserver: Strong ciphers for submission and smtps
We add the stronger ciphers for submission and SMTPS to impove security
for own users, while keep weaker ciphers by default to allow easier mail
transfer.

This benefits security on both ends, as the alternative to a fails TLS
handshake is fallback to plaintext, which is worse than weaker ciphers.