Add ssh support to preboot environment to unlock lucks
This module adds support for SSH to dracut so you can ssh into the 
machine as soon as its booted and unlock luks.

It uses the machines ed25519 key and you can specify a separate or the 
same key to auth to your preboot environment.

Inspired by this repository: https://github.com/gsauthof/dracut-sshd