Install docker-compose as python library from pip
The CentOS package maintainer decided to upgrade docker-compose and 
provide the python3 version only via package. This breaks Ansible that 
relies on using Python2 at least without breaking on CentOS with the 
default python3 version installed.

So installing docker-compose using pip seems to be the way to go and so 
we go it.