traefik: Add docker socket firewall
This change will provide an HA proxy instance that forwards the docker 
socket to traefik. This HA proxy is only available on the internal 
network so no other container is able to talk to docker socket.

Also it'll block all writing requests to the docker socket, which should 
protect the host from a compromised traefik contianer.