static_websites: Provide http alt-svc
With the added support for alt-svc on traefik, we can now push the right 
headers to the applications the torbrowsers can be aware of the hidden 
service address.

This patch should make sure that torbrowser users, now use the hidden 
service to connect all static websites as soon as they visited the the 
clearweb webpage once.