mailserver: Update password algorithm
It's way to long sinc the decision for the password algorithm. This
patch finally gets rid of md5 for the mail passwords and provides
something reasonable secure.

scrypt would still be better, but compared to md5 sha512-crypt is hell a
lot better.

Notice, if you use this role, this is a breaking change, make sure you
update passwords first.