matrix: Rework configs to comply with new standard
Matrix decided to deprecate dynamically generated `homeserver.yaml`s for
their containers. Instead it's recommended to use the generate argument
which will create a `homeserver.yml` in the persistent directory and
allow modifications from there on.

This patch changes the behavior of this role to generate such a config
if it doesn't already exists. It also removes the "SYNAPSE_SERVERNAME"
variable from the `docker-compose.yml` in order to make sure synapse is
picking up the permanent config from the data directory.