mastodon: Fix deployment by using abolute path
To prevent the deployment from breaking on validation of the
docker-compose file, this patch add an absolute patch to the env_file
entry. This should resolve the problem, that the validation command for
the docker-compose file runs in the Ansible runtime directory and
therefore doesn't find the env_file relative to the working directory.